requirements.txt
aiohttp==3.7.3
aiosignal==1.2.0
async-timeout==3.0.1
attrs==20.3.0
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
Flask==2.0.2
frozenlist==1.3.0
idna==3.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
multidict==5.1.0
pyee==9.0.3
python-dotenv==0.15.0
requests==2.27.1
slackclient==2.9.3
slackeventsapi==3.0.1
typing-extensions==3.7.4.3
urllib3==1.26.8
Werkzeug==2.0.2
yarl==1.6.3
